// src/types/enums.ts
export const FurnishingType = {
  UNFURNISHED: "UNFURNISHED",
  SEMI_FURNISHED: "SEMI_FURNISHED",
  FULLY_FURNISHED: "FULLY_FURNISHED",
} as const;

export type FurnishingType =
  (typeof FurnishingType)[keyof typeof FurnishingType];
// ...
export const AvailabilityStatus = {
  READY_TO_MOVE: "READY_TO_MOVE",
  UNDER_CONSTRUCTION: "UNDER_CONSTRUCTION",
  POSSESSION_SOON: "POSSESSION_SOON",
  NEW_LAUNCH: "NEW_LAUNCH",
} as const;

export type AvailabilityStatus =
  (typeof AvailabilityStatus)[keyof typeof AvailabilityStatus];

export const AgeOfProperty = {
  ZERO_TO_ONE: "ZERO_TO_ONE",
  ONE_TO_FIVE: "ONE_TO_FIVE",
  FIVE_TO_TEN: "FIVE_TO_TEN",
  TEN_PLUS: "TEN_PLUS",
} as const;

export type AgeOfProperty = (typeof AgeOfProperty)[keyof typeof AgeOfProperty];
// ...
export const AreaUnit = {
  SQ_FT: "SQ_FT",
  SQ_M: "SQ_M",
  BIGHA: "BIGHA",
  KOTTAH: "KOTTAH",
  GROUNDS: "GROUNDS",
  ARES: "ARES",
  BISWA: "BISWA",
  GUNTHA: "GUNTHA",
  AANKADAM: "AANKADAM",
  HECTARES: "HECTARES",
  CENTS: "CENTS",
  PERCH: "PERCH",
} as const;

export type AreaUnit = (typeof AreaUnit)[keyof typeof AreaUnit];
// ...
export const PropertyOnFloor = {
  LOWER_BASEMENT: "LOWER_BASEMENT",
  BASEMENT: "BASEMENT",
  LOWER_GROUND: "LOWER_GROUND",
  GROUND: "GROUND",
  ROOFTOP_TERRACE: "ROOFTOP_TERRACE",
  FLOOR_1: "FLOOR_1",
  FLOOR_2: "FLOOR_2",
  FLOOR_3: "FLOOR_3",
  FLOOR_4: "FLOOR_4",
  FLOOR_5: "FLOOR_5",
  FLOOR_6: "FLOOR_6",
  FLOOR_7: "FLOOR_7",
  FLOOR_8: "FLOOR_8",
  FLOOR_9: "FLOOR_9",
  FLOOR_10: "FLOOR_10",
  FLOOR_11: "FLOOR_11",
  FLOOR_12: "FLOOR_12",
  FLOOR_13: "FLOOR_13",
  FLOOR_14: "FLOOR_14",
  FLOOR_15: "FLOOR_15",
  FLOOR_16: "FLOOR_16",
  FLOOR_17: "FLOOR_17",
  FLOOR_18: "FLOOR_18",
  FLOOR_19: "FLOOR_19",
  FLOOR_20: "FLOOR_20",
} as const;

export type PropertyOnFloor =
  (typeof PropertyOnFloor)[keyof typeof PropertyOnFloor];